+++ backend/NetboardApi/Dtos/TrafficWindowDtos/TotalDeviceTrafficDto.cs
@@ -4,7 +4,7 @@ public record TotalDeviceTrafficDto
 {
     public string DeviceIp { get; init; } = string.Empty;
     public string Protocol { get; init; } = string.Empty;
-    public int TotalSizeKbps { get; init; }
-    public int UploadSizeKbps { get; init; }
-    public int DownloadSizeKbps { get; init; }
+    public double TotalSizeKbps { get; init; }
+    public double UploadSizeKbps { get; init; }
+    public double DownloadSizeKbps { get; init; }
 }
