+++ backend/NetboardApi/Models/TrafficWindow.cs
@@ -5,9 +5,9 @@ public class TrafficWindow
     public int Id { get; set; }
     public int DeviceId { get; set; }
     public int ProtocolId { get; set; }
-    public int TotalSizeKbps { get; set; }
-    public int UploadSizeKbps { get; set; }
-    public int DownloadSizeKbps { get; set; }
+    public double TotalSizeKbps { get; set; }
+    public double UploadSizeKbps { get; set; }
+    public double DownloadSizeKbps { get; set; }
     public DateTimeOffset CreatedAt { get; set; }
     
     public Protocol Protocol { get; set; } = null!;
